Modify code for compilation in ESP-IDF environment

+++ src/PsychicRequest.cpp
@@ -469,7 +469,7 @@ const String PsychicRequest::_getRandomHexString()
   char buffer[33];  // buffer to hold 32 Hex Digit + /0
   int i;
   for(i = 0; i < 4; i++) {
-    sprintf (buffer + (i*8), "%08lx", esp_random());
+    sprintf (buffer + (i*8), "%08lx", (unsigned long int)esp_random());
   }
   return String(buffer);
 }
